services/autonomyx-developer-agent/requirements.txt
claude-agent-sdk>=0.1.63
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0
aiosqlite>=0.20.0
httpx>=0.27.0
slack-bolt>=1.21.0
